Extract Popularity Lookup from Elasticsearch::Index
